Uploaded picture now getting deleted , fixed form bugs
